+++ CHANGES.md
@@ -1,5 +1,6 @@
 
-
+  - [Eulerian]: Eulerian paths (new module)
+    currently limited to undirected graphs
   - [Components]: strong articulation points (see functors [Connectivity]
     and [BiConnectivity]) (Timothy Bourke)
   - [Dominator]: non-trivial dominators (Timothy Bourke)